requirements.txt
fastapi
uvicorn[standard]
python-dotenv
psycopg2-binary
alembic
Flask
SQLAlchemy
google-generativeai
passlib[bcrypt]
pylint
